+++ src/VSCode.Extension/xaml/completionprovider.ts
